backend_requirements.txt
flask==2.3.3
flask-cors==4.0.0
flask-redis==0.4.0
pandas==2.0.3
sqlite3
google-generativeai==0.3.0
prometheus-client==0.17.1
gunicorn==21.2.0